Use database to store per-file source mode and swap settings.
This also allows thumbnails of videos and regular image files to be stereo if they were previously opened and a stereo mode was set.
Currently per-file swap has no UI, but just defaults to false for .pns & .jps files and true for everything else.
This is in my experience more likely to be accurate than how it swapped previously, which was false for any image and true for video. Still, I will at some point add UI control for the times it isn't accurate.