Fixed Steam Controller button handling when action set changes.
Previously any buttons held down, such as the one that caused the state change, would instantly be recognized by the new action set as being just pressed, and their action would be called.
One example of where this would be a serious problem is binding the button used to open the file browser to an action within the file browser.