Change DVWindow into DVWindowHook, which attaches to a QML window.
The advantage of this is that the window is created in QML, which allows QML to reference it directly.
For example, the fake cursor `parent: root.parent` workaround to get it over popups is no longer needed.
Also, it would be possible to set attached properties (such as style configuration properties) on the window item and have them propogate to all children, popups included.