Remove modes property from plugin manager.
As plugin modes are no longer a thing the only modes available are the built in ones, thus there is no need for a dynamic modes property.
Moved the "list modes" command line argument handling into main() so that the window no longer needs to be inited to run it.