Added bindings & fixed distortion for the Vive.
The OpenVR driver handles lens distortion by default, but leaves the option for programs to take care of it themselves.
The template I originally used had the code to handle distortion in it and it seemed to work fine so I kept it. However, it was actually applying it twice, it's just that the Oculus driver doesn't use OpenVR distortion (I'd assume because the Oculus software does it), so I never noticed (because double nothing is still nothing).
Anyway, now we just let OpenVR take care of the distortion, which cuts out about a hundred lines so that's nice.